requirements.txt
mcp 
mcp[cli]
anthropic
httpx
pyyaml
uvicorn
aiohttp
google-adk
